scheduler/requirements.txt
apache-airflow-providers-singularity==3.8.2
paho-mqtt==2.1.0
